+++ src/lib/components/Link.svelte
@@ -37,7 +37,6 @@
 		font-weight: 400;
 		text-decoration: none;
 		white-space: wrap;
-		align-self: start;
 	}
 
 	a .underline {